refactor: move RaftClusterInstance out of test-common-api: WIP
