+++ test-common/test-common-api/src/main/java/org/apache/kafka/common/test/RaftClusterInstance.java
@@ -1,0 +1,227 @@
+package org.apache.kafka.common.test;
+
+import kafka.network.SocketServer;
+import kafka.server.BrokerServer;
+import kafka.server.ControllerServer;
+import kafka.server.KafkaBroker;
+
+import org.apache.kafka.clients.admin.Admin;
+import org.apache.kafka.common.network.ListenerName;
+import org.apache.kafka.common.test.api.ClusterConfig;
+import org.apache.kafka.common.test.api.ClusterInstance;
+import org.apache.kafka.common.test.api.Type;
+import org.apache.kafka.common.utils.Utils;
+import org.apache.kafka.metadata.BrokerState;
+import org.apache.kafka.metadata.bootstrap.BootstrapMetadata;
+import org.apache.kafka.metadata.storage.FormatterException;
+import org.apache.kafka.server.common.FeatureVersion;
+import org.apache.kafka.server.common.Features;
+import org.apache.kafka.server.common.MetadataVersion;
+
+import java.util.*;
+import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import scala.compat.java8.OptionConverters;
+
+public class RaftClusterInstance implements ClusterInstance {
+
+    private final ClusterConfig clusterConfig;
+    final AtomicBoolean started = new AtomicBoolean(false);
+    final AtomicBoolean stopped = new AtomicBoolean(false);
+    final AtomicBoolean formated = new AtomicBoolean(false);
+    private final ConcurrentLinkedQueue<Admin> admins = new ConcurrentLinkedQueue<>();
+    private KafkaClusterTestKit clusterTestKit;
+    private final boolean isCombined;
+
+    public RaftClusterInstance(ClusterConfig clusterConfig, boolean isCombined) {
+        this.clusterConfig = clusterConfig;
+        this.isCombined = isCombined;
+    }
+
+    @Override
+    public String bootstrapServers() {
+        return clusterTestKit.bootstrapServers();
+    }
+
+    @Override
+    public String bootstrapControllers() {
+        return clusterTestKit.bootstrapControllers();
+    }
+
+    @Override
+    public ListenerName clientListener() {
+        return ListenerName.normalised("EXTERNAL");
+    }
+
+    @Override
+    public Optional<ListenerName> controllerListenerName() {
+        return controllers().values().stream()
+                .findAny()
+                .flatMap(s -> OptionConverters.toJava(s.config().controllerListenerNames().headOption()))
+                .map(ListenerName::new);
+    }
+
+    @Override
+    public Collection<SocketServer> controllerSocketServers() {
+        return controllers().values().stream()
+                .map(ControllerServer::socketServer)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public String clusterId() {
+        return Stream.concat(controllers().values().stream().map(ControllerServer::clusterId),
+                        brokers().values().stream().map(KafkaBroker::clusterId)).findFirst()
+                .orElseThrow(() -> new RuntimeException("No controllers or brokers!"));
+    }
+
+    @Override
+    public Type type() {
+        return isCombined ? Type.CO_KRAFT : Type.KRAFT;
+    }
+
+    @Override
+    public ClusterConfig config() {
+        return clusterConfig;
+    }
+
+    @Override
+    public Set<Integer> controllerIds() {
+        return controllers().keySet();
+    }
+
+    @Override
+    public KafkaClusterTestKit getUnderlying() {
+        return clusterTestKit;
+    }
+
+    @Override
+    public Admin createAdminClient(Properties configOverrides) {
+        Admin admin = Admin.create(clusterTestKit.newClientPropertiesBuilder(configOverrides).build());
+        admins.add(admin);
+        return admin;
+    }
+
+    @Override
+    public void start() {
+        try {
+            format();
+            if (started.compareAndSet(false, true)) {
+                clusterTestKit.startup();
+                TestUtils.waitForCondition(
+                        () -> this.clusterTestKit.brokers().values().stream().allMatch(
+                                brokers -> brokers.brokerState() == BrokerState.RUNNING
+                        ), "Broker never made it to RUNNING state.");
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to start Raft server", e);
+        }
+    }
+
+    @Override
+    public void stop() {
+        if (stopped.compareAndSet(false, true)) {
+            admins.forEach(admin -> Utils.closeQuietly(admin, "admin"));
+            admins.clear();
+            Utils.closeQuietly(clusterTestKit, "cluster");
+        }
+    }
+
+    @Override
+    public void shutdownBroker(int brokerId) {
+        findBrokerOrThrow(brokerId).shutdown();
+    }
+
+    @Override
+    public void startBroker(int brokerId) {
+        findBrokerOrThrow(brokerId).startup();
+    }
+
+    @Override
+    public void waitForReadyBrokers() throws InterruptedException {
+        try {
+            clusterTestKit.waitForReadyBrokers();
+        } catch (ExecutionException e) {
+            throw new AssertionError("Failed while waiting for brokers to become ready", e);
+        }
+    }
+
+    @Override
+    public Map<Integer, KafkaBroker> brokers() {
+        return clusterTestKit.brokers().entrySet()
+                .stream()
+                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+    }
+
+    @Override
+    public Map<Integer, ControllerServer> controllers() {
+        return Collections.unmodifiableMap(clusterTestKit.controllers());
+    }
+
+    public void format() throws Exception {
+        if (formated.compareAndSet(false, true)) {
+            Map<String, Features> nameToSupportedFeature = new TreeMap<>();
+            Features.PRODUCTION_FEATURES.forEach(feature -> nameToSupportedFeature.put(feature.featureName(), feature));
+            Map<String, Short> newFeatureLevels = new TreeMap<>();
+
+            // Verify that all specified features are known to us.
+            for (Map.Entry<Features, Short> entry : clusterConfig.features().entrySet()) {
+                String featureName = entry.getKey().featureName();
+                short level = entry.getValue();
+                if (!featureName.equals(MetadataVersion.FEATURE_NAME)) {
+                    if (!nameToSupportedFeature.containsKey(featureName)) {
+                        throw new FormatterException("Unsupported feature: " + featureName +
+                                ". Supported features are: " + String.join(", ", nameToSupportedFeature.keySet()));
+                    }
+                }
+                newFeatureLevels.put(featureName, level);
+            }
+            newFeatureLevels.put(MetadataVersion.FEATURE_NAME, clusterConfig.metadataVersion().featureLevel());
+
+            // Add default values for features that were not specified.
+            Features.PRODUCTION_FEATURES.forEach(supportedFeature -> {
+                if (!newFeatureLevels.containsKey(supportedFeature.featureName())) {
+                    newFeatureLevels.put(supportedFeature.featureName(),
+                            supportedFeature.defaultValue(clusterConfig.metadataVersion()));
+                }
+            });
+
+            // Verify that the specified features support the given levels. This requires the full
+            // features map since there may be cross-feature dependencies.
+            for (Map.Entry<String, Short> entry : newFeatureLevels.entrySet()) {
+                String featureName = entry.getKey();
+                if (!featureName.equals(MetadataVersion.FEATURE_NAME)) {
+                    short level = entry.getValue();
+                    Features supportedFeature = nameToSupportedFeature.get(featureName);
+                    FeatureVersion featureVersion =
+                            supportedFeature.fromFeatureLevel(level, true);
+                    Features.validateVersion(featureVersion, newFeatureLevels);
+                }
+            }
+
+            TestKitNodes nodes = new TestKitNodes.Builder()
+                    .setBootstrapMetadata(BootstrapMetadata.fromVersions(clusterConfig.metadataVersion(), newFeatureLevels, "testkit"))
+                    .setCombined(isCombined)
+                    .setNumBrokerNodes(clusterConfig.numBrokers())
+                    .setNumDisksPerBroker(clusterConfig.numDisksPerBroker())
+                    .setPerServerProperties(clusterConfig.perServerOverrideProperties())
+                    .setNumControllerNodes(clusterConfig.numControllers()).build();
+            KafkaClusterTestKit.Builder builder = new KafkaClusterTestKit.Builder(nodes);
+            // Copy properties into the TestKit builder
+            clusterConfig.serverProperties().forEach(builder::setConfigProp);
+            // KAFKA-12512 need to pass security protocol and listener name here
+            this.clusterTestKit = builder.build();
+            this.clusterTestKit.format();
+        }
+    }
+
+    private BrokerServer findBrokerOrThrow(int brokerId) {
+        return Optional.ofNullable(clusterTestKit.brokers().get(brokerId))
+                .orElseThrow(() -> new IllegalArgumentException("Unknown brokerId " + brokerId));
+    }
+
+}
