+++ test-common/test-common-api/src/main/java/org/apache/kafka/common/test/api/RaftClusterInvocationContext.java
@@ -1,0 +1,74 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.common.test.api;
+
+
+import org.apache.kafka.common.test.KafkaClusterTestKit;
+import org.apache.kafka.common.test.RaftClusterInstance;
+
+import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
+import org.junit.jupiter.api.extension.BeforeTestExecutionCallback;
+import org.junit.jupiter.api.extension.Extension;
+import org.junit.jupiter.api.extension.TestTemplateInvocationContext;
+
+import java.util.Arrays;
+import java.util.List;
+
+
+/**
+ * Wraps a {@link KafkaClusterTestKit} inside lifecycle methods for a test invocation. Each instance of this
+ * class is provided with a configuration for the cluster.
+ *
+ * This context also provides parameter resolvers for:
+ *
+ * <ul>
+ *     <li>ClusterConfig (the same instance passed to the constructor)</li>
+ *     <li>ClusterInstance (includes methods to expose underlying SocketServer-s)</li>
+ * </ul>
+ */
+public class RaftClusterInvocationContext implements TestTemplateInvocationContext {
+
+    private final String baseDisplayName;
+    private final ClusterConfig clusterConfig;
+    private final boolean isCombined;
+
+    public RaftClusterInvocationContext(String baseDisplayName, ClusterConfig clusterConfig, boolean isCombined) {
+        this.baseDisplayName = baseDisplayName;
+        this.clusterConfig = clusterConfig;
+        this.isCombined = isCombined;
+    }
+
+    @Override
+    public String getDisplayName(int invocationIndex) {
+        return String.format("%s [%d] Type=Raft-%s, %s", baseDisplayName, invocationIndex, isCombined ? "Combined" : "Isolated", String.join(",", clusterConfig.displayTags()));
+    }
+
+    @Override
+    public List<Extension> getAdditionalExtensions() {
+        RaftClusterInstance clusterInstance = new RaftClusterInstance(clusterConfig, isCombined);
+        return Arrays.asList(
+                (BeforeTestExecutionCallback) context -> {
+                    clusterInstance.format();
+                    if (clusterConfig.isAutoStart()) {
+                        clusterInstance.start();
+                    }
+                },
+                (AfterTestExecutionCallback) context -> clusterInstance.stop(),
+                new ClusterInstanceParameterResolver(clusterInstance)
+        );
+    }
+}
